Extract parking lots extraction code into separate file
